Add repo list api endpoint for cli #6941
- moved list method to other resource
